+++ members/nullnet-server/src/db/migrations/2026-08-18-000001_replace_services_with_stack_configs/down.sql
@@ -1,0 +1,7 @@
+DROP TABLE stack_configs;
+
+CREATE TABLE services (
+    stack        TEXT NOT NULL PRIMARY KEY,
+    service_json TEXT NOT NULL,
+    updated_at   BIGINT NOT NULL
+);